+++ Recipes/Rezept - Heißes Wasser.txt
@@ -1,0 +1,7 @@
+Zutaten:
+1L Wasser
+10mL glühend heiße Lava
+
+1) Wasser in einen Topf vorlegen
+2) Die glühend heiße Lava langsam unter schreien in das Wasser geben und stetig rühren
+3) Trinken *Prost*
